backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
Deprecated==1.2.15
easyocr==1.7.2
fastapi==0.115.6
filelock==3.16.1
fire==0.7.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
lazy_loader==0.4
limits==4.0.1
lmdb==1.6.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.15
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.1.0
propcache==0.2.1
protobuf==3.20.2
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic_core==2.27.2
PyMuPDF==1.25.2
pyparsing==3.2.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
